+++ src/index.js
@@ -1,5 +1,7 @@
 /* eslint-disable no-use-before-define */
 
+import { getData } from './modules/tvApi.js';
+import { fetchLikeApi } from './modules/involvementApi.js';
 import './style.css';
 import Logo from './assets/logo.svg';
 
@@ -69,130 +71,3 @@ const displayShow = async () => {
     });
   });
 };
-
-// Likes
-const tvId = 'MjgCDPvMKfBMbwFq4McF';
-const fetchLikeApi = async () => {
-  const getLikeResult = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${tvId}/likes`).then((res) => res.json());
-  return getLikeResult;
-};
-const submitNewLike = async (id) => {
-  await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${tvId}/likes`, {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: id,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-};
-
-const postComments = async (id, name, com) => {
-  await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/4va6c4ouZmpzSETsANV3/comments', {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: id,
-      username: name,
-      comment: com,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-};
-const getComments = async (id) => {
-  const response = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/4va6c4ouZmpzSETsANV3/comments?item_id=${id}`).then((res) => res.json());
-  return response;
-};
-document.addEventListener('click', async (e) => {
-  if (!e.target.matches('.comment-btn')) {
-    return;
-  }
-  const { id } = e.target;
-  const comList = await getComments(id);
-  const showInfo = await getData(id);
-  await display(showInfo, id, comList);
-  const popup = document.querySelector('.popup');
-  popup.classList.add('active');
-  displayCom(comList);
-});
-
-// Popup {
-const getData = async (id) => {
-  const movieInfo = await fetch(`https://api.tvmaze.com/shows/${id}`).then((result) => result.json());
-  return movieInfo;
-};
-const countComments = (number) => number.length;
-const display = async (movieInfo, id, comList) => {
-  const show = await movieInfo;
-  const array = await comList;
-  const popup = document.querySelector('.popup');
-  popup.innerHTML = `
-  <div class="modal-container"> 
-    <div class="modal">
-        <img src="${show.image.original}" alt="${show.name}" class="modal-image">
-        <div>
-          <div class="modal-content">
-            <div class="modal-title">
-              <h1>${show.name}</h1>
-              <a href="#" class="close-modal"><i class="fas fa-times fa-2x"></i></a>
-            </div>
-            <ul class="modal-sub-title">
-              <li>${show.premiered.substring(0, 4)}</li>
-              <li>•</li>
-              <li>${show.language}</li>
-              <li>•</li>
-              <li>${show.averageRuntime} mins</li>
-              <li>•</li>
-              <li><i class="fas fa-star"></i> ${show.rating.average}</li>
-              <li>•</li>
-              <li>${show.status}</li>
-            </ul>
-            <hr>
-            <ul class="genres">
-              <li>${show.genres[0]}</li>
-              <li>${show.genres[1]}</li>
-              <li>${show.genres[2]}</li>
-            </ul>
-            <p>${show.summary}</p>
-            <div class="current-comments">
-              <h4>Comments (${countComments(array)})</h4>
-              <div class="all-comments">
-                <ul class="display-comments"></ul>
-              </div>
-            </div>
-            <div class="add-comments">
-              <h4>Add a comment</h4>
-              <form class="form-add-comment" action="">
-                <input class="input-comment-name" type="text" id="fname" name="fname" placeholder="Your name?" required>
-                <textarea class="input-comment-insight" id="comment" name="comment" placeholder="Your comments?" rows="6" required></textarea>
-                <button id=${id} type="submit">Submit</button>
-              </form>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>`;
-};
-
-const displayCom = async (showInfo) => {
-  const commentList = document.querySelector('.display-comments');
-  commentList.innerHTML = '';
-  const array = await showInfo;
-  array.forEach((item) => {
-    commentList.innerHTML += `<li class="listCom">${item.username}<br>${item.comment}<br>${item.creation_date}</li>`;
-  });
-};
-document.addEventListener('click', async (e) => {
-  if (!e.target.matches('.submit')) {
-    return;
-  }
-  e.preventDefault();
-  const name = document.getElementById('fname').value;
-  const com = document.getElementById('comment').value;
-  const { id } = e.target;
-  await postComments(id, name, com);
-  const comList = await getComments(id);
-  displayCom(comList);
-});
